requirements.txt
streamlit
tensorflow==2.15.0
tf-keras==2.15.0
opencv-python-headless
numpy==1.25.2
pandas
pillow